requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.2.10           # via django
django-cors-headers==3.4.0  # via -r requirements/requirements.in
django-filter==2.3.0      # via -r requirements/requirements.in
django-rest-framework==0.1.0  # via -r requirements/requirements.in
django-templated-mail==1.1.1  # via djoser
django==3.0.8             # via -r requirements/requirements.in, django-cors-headers, django-filter, djangorestframework
djangorestframework==3.11.0  # via django-rest-framework
djoser==2.0.3             # via -r requirements/requirements.in
gunicorn==20.0.4          # via -r requirements/requirements.in
psycopg2-binary==2.8.5    # via -r requirements/requirements.in
pytz==2020.1              # via django
sqlparse==0.3.1           # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools
